Dockerfile: Introduce Docker Image
Add very simple docker file that makes use of the install script. Due to
the install script, the docker file can be used to produce production
images and also images to be used in testing.
Simply add '--build-arg INSTALL_TEST=1' to the docker build command and
get a testing image.

Signed-off-by: Jens Erdmann <jens.erdmann@web.de>